Set focus by range rather than "focus position."
This extends SetFocusPosition to take a range in meters; this is more natural for integration partners and prevents exposure to focus mapping. The 'position' field would continue to be supported by the firmware, but if 'focusRange' is present in the command that will be used instead. ARIScope will continue to use 'position' in support of older firmware.